Updated dates to display month in letters

+++ app/utils/timestampToDate.js
@@ -55,7 +55,7 @@ export const formatDate = (timestamp) => {
 
     return date.toLocaleDateString(undefined, {
       year: "numeric",
-      month: "numeric",
+      month: "short",
       day: "numeric",
     })
   };
